requirements.txt
addcopyfighandler==3.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.2.1
CoolProp==6.6.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.53.0
frozenlist==1.5.0
gurobipy==11.0.3
idna==3.10
joblib==1.4.2
juliacall==0.9.24
juliapkg==0.1.16
kiwisolver==1.4.5
matplotlib==3.10.0
matplotlib-label-lines==0.7.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.25.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.3.6
plotly==6.0.0
ply==3.11
pooch==1.8.2
propcache==0.3.0
pyarrow==19.0.0
pyfluids==2.6.0
Pyomo==6.7.3
pyparsing==3.1.2
pysr==1.5.6
python-dateutil==2.9.0.post0
pytz==2024.1
pyvista==0.44.2
pywin32==306
requests==2.32.3
scikit-learn==1.5.1
scipy==1.13.1
scooby==0.10.0
seaborn==0.13.2
semver==3.0.4
setuptools==80.3.1
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tomli==2.0.1
trame==3.8.0
trame-client==3.6.0
trame-server==3.3.0
trame-vtk==2.8.15
trame-vuetify==2.8.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
vtk==9.3.1
wslink==2.3.2
yarl==1.18.3
